Refactor: Move training utilities to src/utils/artifact_manager
