+++ src/feature_generation/utils/step06_enhanced_feature_engineering_step.py
@@ -47,7 +47,7 @@ except ImportError:
 
 # Import base step and utilities
 # Import BaseStep from training module
-from src.training.base_step import BaseStep
+from src.training.steps.base_step import BaseStep
 from src.utils.comprehensive_function_logger import (
     log_step_functions, log_important_calls, log_all_calls,
     log_internal_call, log_step_progress, log_data_operation
