+++ src/training/utils/__init__.py
@@ -2,6 +2,6 @@
 Training utilities for the Ares framework.
 """
 
-from src.utils.artifact_manager import ArtifactManager, get_analyst_context, setup_enhanced_artifact_manager, get_pretraining_artifact_manager
+from src.utils.artifact_manager import ArtifactManager, get_analyst_context, setup_enhanced_artifact_manager, get_pretraining_artifact_manager, get_step_context_from_config, create_training_artifact_manager, validate_training_config, get_training_metrics, log_training_progress, log_training_error
 
 __all__ = ['ArtifactManager', 'get_analyst_context', 'setup_enhanced_artifact_manager', 'get_pretraining_artifact_manager']
